requirements.txt
amqp==5.3.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
cffi==2.0.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==4.2.27
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
exceptiongroup==1.3.1
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.2
packaging==26.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.0.1
referencing==0.36.2
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.5.3
